add RequestLoggingAspect, ErrorController, tests to Order board. refa…

+++ cookstarter-order-board/src/main/java/ru/guteam/cookstarterorderboard/model/Dish.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Data
@@ -12,5 +13,5 @@ import java.math.BigInteger;
 public class Dish {
     private String name;
     private Integer qty;
-    private BigInteger price;
+    private BigDecimal price;
 }
